requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
click==8.2.1
fastapi==0.116.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
paho-mqtt==2.1.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
